+++ play-messages-library/src/main/scala/com/github/evanbennett/play_messages/Message.scala
@@ -3,14 +3,19 @@
  * All rights reserved.
  */
 
-package com.github.evanbennett.play_messages_scala
+package com.github.evanbennett.play_messages
+
+import play.api.i18n.Lang
 
 /** Message */
 class Message(val key: String) {
+	if (key == null || key.isEmpty) throw new IllegalArgumentException("'key' must be provided.")
 
-	def apply(args: String*)(implicit lang: play.api.i18n.Lang): String = {
-		play.api.i18n.Messages(key, args: _*)
-	}
+	def apply(args: String*)(implicit lang: Lang): String = play.api.i18n.Messages(key, args: _*)
+
+	def get(args: Object*): String = play.i18n.Messages.get(key, args)
+
+	def get(lang: Lang, args: Object*): String = play.i18n.Messages.get(lang, key, args)
 
 	override def equals(a: Any): Boolean = {
 		if (a == null) false
@@ -27,10 +32,7 @@ class Message(val key: String) {
 
 object Message {
 
-	def apply(key: String): Message = {
-		if (key == null || key.isEmpty) throw new IllegalArgumentException("'key' must be provided.")
-		new Message(key)
-	}
+	def apply(key: String): Message = new Message(key)
 
 	def unapply(message: Message): Option[(String)] = {
 		if (message == null) None