+++ sbt-play-messages/src/main/scala/com/github/evanbennett/sbt_play_messages/PlayMessagesPlugin.scala
@@ -7,28 +7,67 @@ package com.github.evanbennett.sbt_play_messages
 
 import sbt._
 
-object PlayMessagesPlugin {
+object PlayMessagesPlugin extends AutoPlugin {
 
-	val BUILD_VERSION = "1.0.0-RC2" // Needs to be in-sync with 'buildVersion' in build.sbt
+	val BUILD_VERSION = "1.1.0-RC1" // Needs to be in-sync with 'buildVersion' in build.sbt
 
-	object PlayMessagesKeys {
-		val requireDefaultMessagesFile = SettingKey[Boolean]("play-messages-require-default-messages-file", "Require a default messages files?")
-		val checkApplicationLanguages = SettingKey[Boolean]("play-messages-check-application-languages", "Check 'application.langs' from 'application.conf' against messages files?")
-		val checkDuplicateKeys = SettingKey[Boolean]("play-messages-check-duplicate-keys", "Check if there are any duplicate keys?")
-		val checkKeyConsistency = SettingKey[Boolean]("play-messages-check-key-consistency", "Check that all the keys match across all messages files?")
-		val checkKeyConsistencySkipFilenames = SettingKey[Set[String]]("play-messages-check-key-consistency-skip-filenames", "Messages filenames to skip when checking for key consistency.")
-		val checkKeysUsed = SettingKey[Boolean]("play-messages-check-keys-used", "Check that all keys are used?")
-		val checkKeysUsedIgnoreKeys = SettingKey[Set[String]]("play-messages-check-keys-used-ignore-keys", "Keys to ignore when checking for use. Regular expressions may be used.")
-		val checkKeysUsedIgnoreFilenames = SettingKey[Set[String]]("play-messages-check-keys-used-ignore-filenames", "Source code filenames to ignore when checking for use.")
-		val generateObject = SettingKey[Boolean]("play-messages-generate-object", "Generate the object with key references?")
-		val generatedObject = SettingKey[String]("play-messages-generated-object", "The object package and name to generate.")
-
-		val checkAndGenerateTask = TaskKey[Seq[File]]("play-messages-check-and-generate-task", "Check the messages and generate the object with key references.")
+	sealed abstract class Status
+	object Statuses {
+		case object Failed extends Status
+		case object Succeeded extends Status
+		case object NoChange extends Status
 	}
 
-	val projectSettings = Seq(
+	object autoImport {
+
+		object PlayMessagesKeys {
+
+			val requireDefaultMessagesFile = SettingKey[Boolean]("play-messages-require-default-messages-file", "Require a default messages files?")
+			val checkApplicationLanguages = SettingKey[Boolean]("play-messages-check-application-languages", "Check 'application.langs' from 'application.conf' against messages files?")
+			val onNoChangeLoadDefaultMessageKeys = SettingKey[Boolean]("play-messages-on-no-change-load-default-message-keys", "In 'checkTask' load (and return) the default message keys even if there is no change. (For use by SBT Plugins extending this one.)")
+			val checkDuplicateKeys = SettingKey[Boolean]("play-messages-check-duplicate-keys", "Check if there are any duplicate keys?")
+			val checkKeyConsistency = SettingKey[Boolean]("play-messages-check-key-consistency", "Check that all the keys match across all messages files?")
+			val checkKeyConsistencySkipFilenames = SettingKey[Set[String]]("play-messages-check-key-consistency-skip-filenames", "Messages filenames to skip when checking for key consistency.")
+			val checkKeysUsed = SettingKey[Boolean]("play-messages-check-keys-used", "Check that all keys are used?")
+			val checkKeysUsedIgnoreKeys = SettingKey[Set[String]]("play-messages-check-keys-used-ignore-keys", "Keys to ignore when checking for use. Regular expressions may be used.")
+			val checkKeysUsedIgnoreFilenames = SettingKey[Set[String]]("play-messages-check-keys-used-ignore-filenames", "Source code filenames to ignore when checking for use.")
+			val generateObject = SettingKey[Boolean]("play-messages-generate-object", "Generate the object with key references?")
+			val generatedObject = SettingKey[String]("play-messages-generated-object", "The object package and name to generate.")
+
+			val checkTask = TaskKey[(Status, Seq[String])]("play-messages-check-task", "Check the messages.")
+			val checkAndGenerateTask = TaskKey[(Status, Seq[File])]("play-messages-check-and-generate-task", "Check the messages and generate the object with key references.")
+		}
+	}
+	import autoImport._
+
+	override lazy val requires = play.Play
+
+	override lazy val trigger = allRequirements
+
+	lazy val checkMessages = Command.command("checkMessages") { state =>
+		Project.runTask(PlayMessagesKeys.checkTask, state) match {
+			case Some((state: State, Value((status: Status, _)))) if status != Statuses.Failed =>
+				if (status == Statuses.NoChange) state.log.info("checkMessages: NO CHANGE")
+				state
+			case Some((state: State, _)) => state.fail
+			case None => state.fail
+		}
+	}
+
+	lazy val generateMessagesObject = Command.command("generateMessagesObject") { state =>
+		Project.runTask(PlayMessagesKeys.checkAndGenerateTask, state) match {
+			case Some((state: State, Value((status: Status, _)))) if status != Statuses.Failed =>
+				if (status == Statuses.NoChange) state.log.info("generateMessagesObject: NO CHANGE")
+				state
+			case Some((state: State, _)) => state.fail
+			case None => state.fail
+		}
+	}
+
+	override lazy val projectSettings = Seq(
 		PlayMessagesKeys.requireDefaultMessagesFile := true,
 		PlayMessagesKeys.checkApplicationLanguages := true,
+		PlayMessagesKeys.onNoChangeLoadDefaultMessageKeys := false,
 		PlayMessagesKeys.checkDuplicateKeys := true,
 		PlayMessagesKeys.checkKeyConsistency := true,
 		PlayMessagesKeys.checkKeyConsistencySkipFilenames := Set.empty[String],
@@ -38,38 +77,32 @@ object PlayMessagesPlugin {
 		PlayMessagesKeys.generateObject := true,
 		PlayMessagesKeys.generatedObject := "conf.Messages",
 
-		Keys.sourceGenerators in Compile <+= PlayMessagesKeys.checkAndGenerateTask
+		PlayMessagesKeys.checkTask := PlayMessages.checkTask.value,
+
+		Keys.libraryDependencies += "com.github.evanbennett" %% "play-messages-library" % BUILD_VERSION,
+
+		Keys.sourceGenerators in Compile <+= Def.task(PlayMessagesKeys.checkAndGenerateTask.value._2),
+
+		Keys.commands ++= Seq(checkMessages, generateMessagesObject)
 	)
 }
 
 object PlayMessagesPluginScala extends AutoPlugin {
-
-	object autoImport {
-		val PlayMessagesKeys = PlayMessagesPlugin.PlayMessagesKeys
-	}
+	import PlayMessagesPlugin.autoImport._
 
 	override lazy val requires = play.PlayScala
 
 	override lazy val trigger = allRequirements
 
-	override lazy val projectSettings = PlayMessagesPlugin.projectSettings ++ Seq(
-		PlayMessagesPlugin.PlayMessagesKeys.checkAndGenerateTask := PlayMessages.checkAndGenerateScalaTask.value,
-
-		Keys.libraryDependencies += "com.github.evanbennett" %% "play-messages-scala" % PlayMessagesPlugin.BUILD_VERSION
-	)
+	override lazy val projectSettings = Seq(PlayMessagesKeys.checkAndGenerateTask := PlayMessages.checkAndGenerateScalaTask.value)
 }
 
 object PlayMessagesPluginJava extends AutoPlugin {
-
-	// 'object autoImport' is not needed as it exists in PlayMessagesPluginScala
+	import PlayMessagesPlugin.autoImport._
 
 	override lazy val requires = play.PlayJava
 
 	override lazy val trigger = allRequirements
 
-	override lazy val projectSettings = PlayMessagesPlugin.projectSettings ++ Seq(
-		PlayMessagesPlugin.PlayMessagesKeys.checkAndGenerateTask := PlayMessages.checkAndGenerateJavaTask.value,
-
-		Keys.libraryDependencies += "com.github.evanbennett" %% "play-messages-java" % PlayMessagesPlugin.BUILD_VERSION
-	)
+	override lazy val projectSettings = Seq(PlayMessagesKeys.checkAndGenerateTask := PlayMessages.checkAndGenerateJavaTask.value)
 }